refactor account related endpoints
