+++ app/index.js
@@ -3,6 +3,7 @@ const express = require('express');
 const cookieParser = require('cookie-parser');
 const cors = require('cors');
 
+const account = require('./account.js');
 const users = require('./users.js');
 const events = require('./events.js');
 const friends = require('./friends.js');
@@ -35,6 +36,7 @@ app.get(`/test`, (req,res) => {
   res.status(200).send("yay!");
 })
 
+app.use('/account', account);
 app.use('/users', users);
 app.use('/events', events);
 app.use('/friends',friends);
